requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-limiter>=3.5.0
flask-caching>=2.1.0
libsass>=0.23.0
watchdog>=3.0.0
psutil>=5.9.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0